core/requirements/dev.txt
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
coverage==4.4.1
Django==1.11.5
django-debug-toolbar==1.8
django-extensions==1.8.1
djangorestframework==3.6.4
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
pbr==3.1.1
psycopg2==2.7.3
py==1.4.34
pytest==3.1.3
pytest-cov==2.5.1
pytz==2017.2
requests==2.18.2
six==1.10.0
sqlparse==0.2.3
stevedore==1.25.0
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
